requirements.txt
Django
djangorestframework
ipython
gunicorn
pandas
openpyxl
django-cors-headers
psycopg2-binary
snakemake
PyVCF3
pysam
matplotlib
numpy
celery
redis
pyarrow